requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.3.0
torch>=2.2.0
numpy>=1.26.0
opencv-python>=4.9.0
pyyaml>=6.0
matplotlib>=3.8.0
pandas>=2.2.0
